requirements.txt
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
Flask==3.0.0
Flask-CORS==4.0.0